# backend/src/swagger.json
{
   "openapi": "3.0.0",
   "info": {
      "title": "Space App",
      "description": "Final fullstack project for Okten School",
      "version": "1.0.0"
   },
   "servers": [
      {
         "url": "http://localhost:5400/api",
         "description": "Express"
      }
   ],
   "paths": {
...
      "/moments": {
         "get": {
            "summary": "get all moments",
            "tags": [
               "Moments"
            ],
            "operationId": "getAllMoments",
            "parameters": [
               {
                  "in": "query",
                  "name": "searchKey",
                  "schema": {
                     "type": "string"
                  },
                  "description": "Keyword for searching"
               }
            ],
            "security": [
               {
                  "Authorization": []
               }
            ],
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Moments"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/moments/add": {
         "get": {
            "summary": "add moment",
            "tags": [
               "Moments"
            ],
            "operationId": "addMoment",
            "security": [
               {
                  "Authorization": []
               }
            ],
            "responses": {
               "201": {
                  "description": "New moment was created",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Moment"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/moments/{momentId}": {
         "get": {
            "summary": "get moment by id",
            "tags": [
               "Moments"
            ],
            "operationId": "getMomentById",
            "parameters": [
               {
                  "name": "momentId",
                  "in": "path",
                  "required": true,
                  "description": "moment id",
                  "schema": {
                     "type": "string"
                  },
                  "example": "63dfe16eda233c96fc6e2604"
               }
            ],
            "security": [
               {
                  "Authorization": []
               }
            ],
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Moment"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ObjectID",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ObjectIdError"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/NotExistError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         },
         "patch": {
            "summary": "update moment by id",
            "tags": [
               "Moments"
            ],
            "operationId": "updateMomentById",
            "parameters": [
               {
                  "name": "momentId",
                  "in": "path",
                  "required": true,
                  "description": "moment id",
                  "schema": {
                     "type": "string"
                  },
                  "example": "63dfe16eda233c96fc6e2604"
               }
            ],
            "security": [
               {
                  "Authorization": []
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "title": {
                              "type": "string",
                              "example": "Dubai morning"
                           },
                           "date": {
                              "type": "number",
                              "example": 1677517089217
                           },
                           "location": {
                              "type": "string",
                              "example": "Dubai"
                           },
                           "tags": {
                              "type": "array",
                              "items": {
                                 "type": "string",
                                 "example": "family"
                              }
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Success"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ObjectID",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ObjectIdError"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/NotExistError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         },
         "delete": {
            "summary": "delete moment by id",
            "tags": [
               "Moments"
            ],
            "operationId": "deleteMomentById",
            "parameters": [
               {
                  "name": "momentId",
                  "in": "path",
                  "required": true,
                  "description": "moment id",
                  "schema": {
                     "type": "string"
                  },
                  "example": "63dfe16eda233c96fc6e2604"
               }
            ],
            "security": [
               {
                  "Authorization": []
               }
            ],
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Success"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ObjectID",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ObjectIdError"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/NotExistError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/moments/{momentId}/photo_upload": {
         "patch": {
            "summary": "upload photo",
            "tags": [
               "Moments"
            ],
            "operationId": "uploadPhoto",
            "parameters": [
               {
                  "name": "momentId",
                  "in": "path",
                  "required": true,
                  "description": "moment id",
                  "schema": {
                     "type": "string"
                  },
                  "example": "63dfe16eda233c96fc6e2604"
               }
            ],
            "security": [
               {
                  "Authorization": []
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "multipart/form-data": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "photo": {
                              "type": "string",
                              "format": "base64"
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "image": {
                                 "type": "string",
                                 "example": "1677521282980.jpg"
                              }
                           }
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ObjectID",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ObjectIdError"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/NotExistError"
                        }
                     }
                  }
               },
               "413": {
                  "description": "File size error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "statusCode": {
                                 "type": "number",
                                 "example": 413
                              },
                              "message": {
                                 "type": "string",
                                 "example": "File size must be less than 3 mb"
                              }
                           }
                        }
                     }
                  }
               },
               "422": {
                  "description": "File type error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "statusCode": {
                                 "type": "number",
                                 "example": 422
                              },
                              "message": {
                                 "type": "string",
                                 "example": "Invalid file type"
                              }
                           }
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      }
   },
   "components": {
      "securitySchemes": {
         "Authorization": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
         }
      },
      "schemas": {
         "UnauthorizedError": {
            "type": "object",
            "required": [
               "statusCode",
               "message"
            ],
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 401
               },
               "message": {
                  "type": "string",
                  "example": "Invalid or expired token"
               }
            }
         },
         "ObjectIdError": {
            "type": "object",
            "required": [
               "statusCode",
               "message"
            ],
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 400
               },
               "message": {
                  "type": "string",
                  "example": "Object ID is not valid"
               }
            }
         },
         "NotExistError": {
            "type": "object",
            "required": [
               "statusCode",
               "message"
            ],
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 404
               },
               "message": {
                  "type": "string",
                  "example": "Object does not exist"
               }
            }
         },
         "DefaultError": {
            "type": "object",
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 500
               },
               "message": {
                  "type": "string",
                  "example": "Error message"
               }
            }
         },
         "Success": {
            "type": "object",
            "properties": {
               "message": {
                  "type": "string",
                  "example": "Success"
               }
            }
         },
...
         "Moment": {
            "type": "object",
            "properties": {
               "id": {
                  "type": "string",
                  "example": "63dfe16eda233c96fc6e2604"
               },
               "title": {
                  "type": "string",
                  "example": "My birthday"
               },
               "photo": {
                  "type": "string",
                  "example": "3820175937533.jpg"
               },
               "date": {
                  "type": "number",
                  "example": 1677517089217
               },
               "location": {
                  "type": "string",
                  "example": "Location"
               },
               "createdAt": {
                  "type": "number",
                  "example": 1677517089217
               },
               "tags": {
                  "type": "array",
                  "items": {
                     "type": "string",
                     "example": "tag"
                  }
               }
            }
         },
         "Moments": {
            "type": "object",
            "properties": {
               "data": {
                  "type": "array",
                  "items": {
                     "$ref": "#/components/schemas/Moment"
                  }
               },
               "tagsForFilter": {
                  "type": "array",
                  "items": {
                     "type": "string",
                     "example": "tags"
                  }
               }
            }
         },
...
      }
   }
}